First build of the basic board mechanics.
Board is completed, capable of storing squares and units and providing
basic interface with squares such as accessing adjacent squares.

Square is capable of handling units properly.

SquareSet is currently appearing to be functioning properly, storing only
single copies of Squares being added to it and ignoring 'null' squares

First round of unit testing finished as well. So far so good.