Improved Pathing algorithm, added documentation to older classes and version numbers.
One thing to keep in mind, be VERY CAREFUL when writing your test cases.
Errors in test cases are hard to catch. Especially when adding multiple
units onto the board, check to make sure that you are performing the
correct methods onto the correct units. It is easy to mess up with a lot
of copy and paste.